Moving files, naming, format
formatted some code according rules for code formatting, moved to folder `common` and merge master.
