+++ common/include/constants.hpp
@@ -4,5 +4,6 @@
 #include <string>
 
 static const std::string SOCKET_NAME{"unix:/tmp/execution_management"};
+static const std::string APPLICATIONS_PATH{"./bin/applications/"};
 
 #endif // CONSTANTS
