+++ common/include/i_application_handler.hpp
