+++ common/include/i_manifest_reader.hpp
